requirements.txt
streamlit
st-pages==0.5.0
streamlit-js-eval
ibm-generative-ai==2.3.0
tqdm
python-calamine
chardet
ibm_watsonx_ai==1.0.9
scipy
matplotlib
python-dotenv
# wikipediaapi # for downloading wiki pages
# openpyxl # for downloading wiki pages
